services/resume_service/requirements.txt
fastapi
uvicorn
python-docx
reportlab
pydantic
